fix: 修复 MaaWpfGui 编译警告
- CS8600: RainbowAnimationBehavior.cs 中 null 字面量转 Brush 类型显式标注
- SA1210: AchievementTrackerHelper.cs 中 using 指令按命名空间字母序重排
- SA1649/SA1402: TolerantEnumConverterFactory 拆分至独立文件
- SA1508: AchievementListDialogView.xaml.cs 删除闭合括号前多余空行
- SA1501: SettingsViewModel.cs 中单行 if 语句展开为多行
