+++ src/MaaWpfGui/Helper/AchievementTrackerHelper.cs
@@ -16,13 +16,13 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Media;
-using System.Windows.Threading;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Threading;
 using HandyControl.Controls;
 using HandyControl.Data;
 using HandyControl.Tools.Command;
