+++ src/MaaWpfGui/Views/Dialogs/AchievementListDialogView.xaml.cs
@@ -33,5 +33,4 @@ public partial class AchievementListDialogView
         // 关闭窗口时执行一次空搜索，重置可见性
         AchievementTrackerHelper.Instance.Search();
     }
-
 }
